making authorized request

+++ frontend/src/components/WorkoutDetails.js
@@ -1,18 +1,27 @@
 import React from 'react'
 import useWorkoutsContext from '../hooks/useWorkoutsContext'
+import useAuthContext from '../hooks/useAuthContext'
 
 // date-fns
 import { formatDistanceToNow } from 'date-fns'
 export default function WorkoutDetails({ workout }) {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const res = await fetch(`/api/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+
+            }
         })
         const data = await res.json()
         if (res.ok) {
-            console.log(data)
+            // console.log(data)
             dispatch({ type: 'DELETE_WORKOUT', payload: data.workout })
         }
     }
